backend/requirements.txt
# Cafeteria CLEO - Backend Dependencies
# Requiere Python 3.11 o 3.12 (3.14 aún no tiene wheels para pydantic-core)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.0
psycopg2-binary>=2.9.9
pydantic>=2.5.0,<3
pydantic-settings>=2.1.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.1,<4.1
python-multipart==0.0.9
email-validator==2.1.0.post1
jinja2==3.1.3
aiosmtplib==3.0.1
httpx==0.26.0
python-dotenv==1.0.1
reportlab==4.0.9
openpyxl==3.1.2
qrcode[pil]==7.4.2
